Make safe_freeze_time a more generic helper
